forge-app/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.0.0",
    "anthropic>=0.30.0",
    "arxiv>=2.1.3",
    "boto3>=1.42.59",
    "botocore>=1.42.59",
    "cerebras-cloud-sdk>=1.0.0",
    "ddgs>=9.11.1",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.111.0",
    "openai>=2.24.0",
    "parallel-web>=0.0.1",
    "pydantic>=2.7.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
]
